requirements.txt
dialogflow==0.7.2
python-telegram-bot==12.3.0
requests==2.22.0
vk-api==11.7.0
